feature-store/server/requirements.txt
aiohttp==3.7.4
psycopg2-binary